+++ package.json
@@ -18,12 +18,14 @@
   "devDependencies": {
     "@babel/core": "7.21.4",
     "@babel/preset-env": "7.21.4",
+    "css-loader": "^7.1.2",
     "babel-loader": "9.1.2",
     "copy-webpack-plugin": "11.0.0",
     "eslint": "8.38.0",
     "eslint-config-htmlacademy": "9.0.0",
     "html-webpack-plugin": "5.5.1",
     "webpack": "5.79.0",
+    "style-loader": "^4.0.0",
     "webpack-cli": "5.0.1",
     "webpack-dev-server": "4.13.3"
   },
