+++ src/presenter/main-presenter.js
@@ -1,42 +1,91 @@
-import { render } from '../render';
-import EditFormView from '../view/edit-form-view';
+import { render, RenderPosition } from '../render';
 import FiltersView from '../view/filters-view';
+import SortingView from '../view/sorting-view';
 import RoutePointView from '../view/route-point-view';
 import RoutePointListView from '../view/route-point-list-view';
-import SortingView from '../view/sorting-view';
+import EditFormView from '../view/edit-form-view';
+import ListEmptyView from '../view/list-empty-view';
 
 export default class MainPresenter {
-  RoutePointListComponent = new RoutePointListView();
+  #routePointListComponent = new RoutePointListView();
+  #routePointsComponents = new Map();
 
-  constructor(DestinationModel, EventModel, OfferModel) {
+  constructor(destinationModel, eventModel, offerModel) {
     this.tripEvents = document.querySelector('.trip-events');
     this.tripControlFilters = document.querySelector('.trip-controls__filters');
-    this.destinationModel = DestinationModel;
-    this.eventModel = EventModel;
-    this.offerModel = OfferModel;
+    this.destinationModel = destinationModel;
+    this.eventModel = eventModel;
+    this.offerModel = offerModel;
   }
 
   init() {
     const events = this.eventModel.getEvents();
 
-    render(new FiltersView(), this.tripControlFilters);
+    render(new FiltersView(this.#generateFilters(events)), this.tripControlFilters);
     render(new SortingView(), this.tripEvents);
-    render(this.RoutePointListComponent, this.tripEvents);
-    render(
-      new EditFormView(events[0], this.destinationModel, this.offerModel),
-      this.RoutePointListComponent.getElement()
+
+    if (events.length === 0) {
+      render(new ListEmptyView(), this.tripEvents);
+      return;
+    }
+
+    render(this.#routePointListComponent, this.tripEvents);
+    for (let i = 0; i < Math.min(3, events.length); i++) {
+      this.#renderRoutePoint(events[i]);
+    }
+
+    document.addEventListener('keydown', this.#handleEscapeKey);
+  }
+
+  #generateFilters(events) {
+    return [
+      { name: 'everything', isDisabled: events.length === 0 },
+      { name: 'future', isDisabled: !events.some((event) => event.dateFrom > Date.now()) },
+      { name: 'past', isDisabled: !events.some((event) => event.dateTo < Date.now()) },
+      { name: 'present', isDisabled: !events.some((event) => event.dateTo === Date.now()) },
+    ];
+  }
+
+  #renderRoutePoint(event) {
+    const destination = this.destinationModel.getDestinationById(event.destination);
+    const eventOffers = this.offerModel.getOffersByType(event.type).filter((offer) =>
+      event.offers.includes(offer.id)
     );
 
-    for (let i = 0; i < Math.min(3, events.length); i++) {
-      const event = events[i];
-      const destination = this.destinationModel.getDestinationById(event.destination);
-      const eventOffers = this.offerModel.getOffersByType(event.type).filter((offer) =>
-        event.offers.includes(offer.id)
-      );
-      render(
-        new RoutePointView(event, destination, eventOffers),
-        this.RoutePointListComponent.getElement()
-      );
-    }
+    const routePointComponent = new RoutePointView(event, destination, eventOffers);
+    const editFormComponent = new EditFormView(event, this.destinationModel, this.offerModel);
+
+    routePointComponent.setEditClickHandler(() =>
+      this.#replaceRoutePointToEditForm(routePointComponent, editFormComponent)
+    );
+    editFormComponent.setFormSubmitHandler(() =>
+      this.#replaceEditFormToRoutePoint(routePointComponent, editFormComponent)
+    );
+    editFormComponent.setRollupClickHandler(() =>
+      this.#replaceEditFormToRoutePoint(routePointComponent, editFormComponent)
+    );
+
+    this.#routePointsComponents.set(routePointComponent, editFormComponent);
+    render(routePointComponent, this.#routePointListComponent.element, RenderPosition.BEFOREEND);
+    routePointComponent.addEventListeners();
   }
+
+  #replaceRoutePointToEditForm(routePointComponent, editFormComponent) {
+    this.#routePointListComponent.element.replaceChild(editFormComponent.element, routePointComponent.element);
+  }
+
+  #replaceEditFormToRoutePoint(routePointComponent, editFormComponent) {
+    this.#routePointListComponent.element.replaceChild(routePointComponent.element, editFormComponent.element);
+  }
+
+  #handleEscapeKey = (evt) => {
+    if (evt.key === 'Escape') {
+      for (const [routePointComponent, editFormComponent] of this.#routePointsComponents) {
+        if (this.#routePointListComponent.element.contains(editFormComponent.element)) {
+          this.#replaceEditFormToRoutePoint(routePointComponent, editFormComponent);
+          break;
+        }
+      }
+    }
+  };
 }
