+++ src/render.js
@@ -13,7 +13,7 @@ function createElement(template) {
 }
 
 function render(component, container, place = RenderPosition.BEFOREEND) {
-  container.insertAdjacentElement(place, component.getElement());
+  container.insertAdjacentElement(place, component.element);
 }
 
 export {RenderPosition, createElement, render};
