+++ src/view/create-form-view.js
@@ -1,19 +1,12 @@
-import {createElement} from '../render.js';
+import AbstractView from '../framework/view/abstract-view.js';
 
-export default class CreateFormView {
-  getTemplate(){
+export default class CreateFormView extends AbstractView {
+  constructor() {
+    super();
+  }
+
+  get template() {
     return createFormTemplate();
-  }
-
-  getElement() {
-    if(!this.element){
-      this.element = createElement(this.getTemplate());
-    }
-    return this.element;
-  }
-
-  removeElement(){
-    this.element = null;
   }
 }
