+++ src/view/filters-view.js
@@ -1,69 +1,25 @@
-import {createElement} from '../render.js';
+import AbstractView from '../framework/view/abstract-view.js';
 
-export default class FiltersView {
-  getTemplate(){
-    return createFiltersTemplate();
+export default class FiltersView extends AbstractView {
+  constructor(filters) {
+    super();
+    this.filters = filters;
   }
 
-  getElement(){
-    if(!this.element){
-      this.element = createElement(this.getTemplate());
-    }
-    return this.element;
+  get template(){
+    return this.createFiltersTemplate();
   }
 
-  removeElement(){
-    this.element = null;
+  createFiltersTemplate() {
+    return `
+    <form class="trip-filters" action="#" method="get">
+      ${this.filters.map(({ name, isDisabled }) => `
+        <div class="trip-filters__filter">
+          <input id="filter-${name}" class="trip-filters__filter-input visually-hidden" type="radio" name="trip-filter" value="${name}" ${isDisabled ? 'disabled' : ''}>
+          <label class="trip-filters__filter-label" for="filter-${name}">${name}</label>
+        </div>
+      `).join('')}
+    </form>
+  `;
   }
-}
-
-function createFiltersTemplate(){
-  return `<form class="trip-filters" action="#" method="get">
-            <div class="trip-filters__filter">
-              <input
-                id="filter-everything"
-                class="trip-filters__filter-input visually-hidden"
-                type="radio"
-                name="trip-filter"
-                value="everything"
-              >
-              <label class="trip-filters__filter-label" for="filter-everything">Everything</label>
-            </div>
-
-            <div class="trip-filters__filter">
-              <input
-                id="filter-future"
-                class="trip-filters__filter-input visually-hidden"
-                type="radio"
-                name="trip-filter"
-                value="future"
-              >
-              <label class="trip-filters__filter-label" for="filter-future">Future</label>
-            </div>
-
-            <div class="trip-filters__filter">
-              <input
-                id="filter-present"
-                class="trip-filters__filter-input visually-hidden"
-                type="radio"
-                name="trip-filter"
-                value="present"
-              >
-              <label class="trip-filters__filter-label" for="filter-present">Present</label>
-            </div>
-
-            <div class="trip-filters__filter">
-              <input
-                id="filter-past"
-                class="trip-filters__filter-input visually-hidden"
-                type="radio"
-                name="trip-filter"
-                value="past"
-                checked
-              >
-              <label class="trip-filters__filter-label" for="filter-past">Past</label>
-            </div>
-
-            <button class="visually-hidden" type="submit">Accept filter</button>
-          </form>`;
 }
