+++ src/view/route-point-list-view.js
@@ -1,20 +1,12 @@
-import { createElement } from '../render.js';
+import AbstractView from '../framework/view/abstract-view.js';
 
-export default class RoutePointListView {
-  getTemplate() {
+export default class RoutePointListView extends AbstractView {
+  constructor() {
+    super();
+  }
+
+  get template() {
     return createPointListTemplate();
-  }
-
-  getElement() {
-    if (!this.element) {
-      this.element = createElement(this.getTemplate());
-    }
-
-    return this.element;
-  }
-
-  removeElement() {
-    this.element = null;
   }
 }
