+++ src/view/sorting-view.js
@@ -1,19 +1,12 @@
-import {createElement} from '../render.js';
+import AbstractView from '../framework/view/abstract-view.js';
 
-export default class SortingView {
-  getTemplate(){
+export default class SortingView extends AbstractView {
+  constructor() {
+    super();
+  }
+
+  get template(){
     return createSortTemplate();
-  }
-
-  getElement(){
-    if(!this.element){
-      this.element = createElement(this.getTemplate());
-    }
-    return this.element;
-  }
-
-  removeElement(){
-    this.element = null;
   }
 }
